Restructure to work with venv and entry points
