+++ .gitignore
@@ -1,1 +1,4 @@
 .api_key
+python
+**/*.egg-info/*
+**/__pycache__/*
