+++ README.md
@@ -1,0 +1,24 @@
+# Weather Seeker
+
+## Quick Start
+1. Create a Python 3 virtual environment for `weather`
+    ```
+    python3 -m venv python
+    source python/bin/activate
+    pip install --upgrade pip setuptools
+    ```
+
+2. Install Weather Seeker
+    ```
+    pip install -e .
+    ```
+
+3. Run `weather`
+    ```
+    weather
+    ```
+
+4. Deactivate when done
+    ```
+    deactivate
+    ```
