+++ src/weather/__init__.py
@@ -1,0 +1,2 @@
+from .__main__ import main
+from .weather import WeatherSeeker
